pyproject.toml
[tool.poetry]
name = "aiohttp_slacklogin"
version = "0.1.0"
description = ""
authors = ["Atsushi Odagiri <aodagx@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.5"
aiohttp_jinja2 = "^1.1"
aiohttp_session = "^2.7"
aiohttp_security = "^0.4.0"
typing_extensions = "^3.7"

[tool.poetry.dev-dependencies]
black = "=18.9b0"
flake8 = "^3.7"
mypy = "^0.670.0"
invoke = "^1.2"
python-dotenv = "^0.10.1"
cryptography = "^2.6"
pytest = "^4.3"
pytest-cov = "^2.6"
pytest-randomly = "^2.1"
pytest-mock = "^1.10"
isort = "^4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
multi_line_output = 3